move fixtures to conftest.py
