+++ test/test_rm.py
@@ -1,11 +1,12 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
-from os.path import join
+from os.path import dirname, join
 import shutil
 
 from asv import config
 from asv import results
 
 from . import tools
+from .tools import example_results
 
 
 def test_rm(tmpdir, example_results):
@@ -17,7 +18,8 @@ def test_rm(tmpdir, example_results):
 
     conf = config.Config.from_json({
         'results_dir': join(tmpdir, 'example_results'),
-        'repo': "### IGNORED, BUT REQUIRED ###"})
+        'repo': "### IGNORED, BUT REQUIRED ###"
+        })
 
     tools.run_asv_with_conf(conf, 'rm', '-y', 'benchmark=time_quantity*')
